# Conflicts:
#	logix.js

+++ logix.js
@@ -175,6 +175,8 @@ function setup() {
     simulationArea.setup();
     scheduleUpdate();
 
+    combinationalAnalysis();
+
 
 }
 
@@ -507,6 +509,7 @@ function CircuitElement(x, y, scope, dir, bitWidth) {
     this.objectType = this.constructor.name; // CHECK IF THIS IS VALID
     this.x = x;
     this.y = y;
+
     this.hover = false;
     if(this.x==undefined||this.y==undefined){
         this.x=simulationArea.mouseX;
@@ -514,6 +517,8 @@ function CircuitElement(x, y, scope, dir, bitWidth) {
         this.newElement=true;
         this.hover=true;
     }
+    this.deleteNodesWhenDeleted=false;
+
     this.parent = parent;
     this.nodeList = []
     this.clicked = false;
@@ -821,11 +826,18 @@ function CircuitElement(x, y, scope, dir, bitWidth) {
     this.delete = function() {
         simulationArea.lastSelected = undefined;
         this.scope[this.objectType].clean(this); // CHECK IF THIS IS VALID
-        for (var i = 0; i < this.nodeList.length; i++) {
-            this.nodeList[i].delete();
-        }
+        if(this.deleteNodesWhenDeleted)
+            for (var i = 0; i < this.nodeList.length; i++)
+                this.nodeList[i].delete();
+        else
+            for (var i = 0; i < this.nodeList.length; i++)
+                this.nodeList[i].converToIntermediate();
     }
 
+    this.cleanDelete=function(){
+        this.deleteNodesWhenDeleted=true;
+        this.delete();
+    }
     //method to change direction
     //OVERRIDE WITH CAUTION
     this.newDirection = function(dir) {
